requirements.txt
Django==1.5
PIL==1.1.7
distribute==0.6.24
dj-database-url==0.2.1
django-appconf==0.6
django-compressor==1.2
django-friendship==0.8.1
django-gravatar==0.1.0
django-guardian==1.1.0.beta
django-registration==0.9
django-userena==1.2.0
easy-thumbnails==1.2
psycopg2==2.4.6
six==1.2.0
wsgiref==0.1.2
